extract NameFilterTest from FilterTest
