+++ src/test/java/spoon/test/filters/FilterTest.java
@@ -19,7 +19,6 @@ package spoon.test.filters;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import spoon.Launcher;
-import spoon.legacy.NameFilter;
 import spoon.reflect.code.CtCFlowBreak;
 import spoon.reflect.code.CtExpression;
 import spoon.reflect.code.CtFieldAccess;
@@ -107,15 +106,6 @@ public class FilterTest {
 	@BeforeEach
 	public void setup() throws Exception {
 		factory = build(Foo.class);
-	}
-
-	@Test
-	public void testNameFilter() throws Exception {
-		// contract: legacy NameFilter is tested and works
-		CtClass<?> foo = factory.Package().get("spoon.test.filters.testclasses").getType("Foo");
-		assertEquals("Foo", foo.getSimpleName());
-		List<CtNamedElement> elements = foo.getElements(new NameFilter<>("i"));
-		assertEquals(1, elements.size());
 	}
 
 	@Test
